applied styling to various components
