+++ client/src/components/PostPreview/PostPreview.css
@@ -5,7 +5,7 @@
     border: 2px solid black;
     padding: 1rem 1rem 0 1rem;
     background-color: #fff;
-    margin: 1.5rem 0;
+    margin: 1.5rem 1rem;
 }
 .post-preview-header {
     display: flex;
