+++ client/src/features/index.ts
@@ -1,5 +1,5 @@
 import NavigationBar from './NavigationBar/NavigationBar'
 import Post from './Post'
-import PostCreator from './PostCreator'
+import PostCreator from './PostCreator/PostCreator'
 
 export { NavigationBar, Post, PostCreator }
