+++ client/src/pages/NewPostPage.tsx
@@ -2,8 +2,8 @@ import { PostCreator } from '../features'
 
 export default function NewPostPage() {
   return (
-    <>
-    <PostCreator />
-    </>
+    <div className="centered-container">
+      <PostCreator />
+    </div>
   )
 }
