+++ client/src/pages/Profile.tsx
@@ -9,7 +9,7 @@ export default function Profile() {
     <div className="profile">
       <img className="profile-avatar" src={dummyData.avatar}></img>
       <button>Edit Avatar</button>
-      <div>{dummyData.username}</div>
+      <div className="profile-username">{dummyData.username}</div>
       <button>Edit Username</button>
       <button>Log Out</button>
     </div>
